Vehicle_Vweup - fix a compilation issue
The compiler in ESP-IDF v5+ is very sensitive to detecting potential issue.
Here, an indentation could be interpreted by the human reader as if an `if` block was encompassing a whole indented block.

The fix here is either to remove the indentation, or to add curly braces to explicitely delimit the block managed by the `if` condition.

The indentation was removed.